feat (activity): 승패기록하는 기능추가

+++ app/src/main/java/woowacourse/omok/data/Player.kt
@@ -4,7 +4,7 @@ import androidx.annotation.DrawableRes
 import java.io.Serializable
 
 data class Player(
-    val id: Int = -1,
+    val id: Long = -1,
     val name: String,
     val win: Int = 0,
     val lose: Int = 0,
